+++ src/shared/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useImageResizer } from './useImageResizer';
 export { useImageProcessingToast } from './useImageProcessingToast';
+export { useAuthImageSize } from './useAuthImageSize';
